==> src/5_Figures.Rmd <==
---
title: "Figures"
author: "Merritt Harlan"
date: '2022-10-18'
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

# Figure 1

Figure 1 was created using QGIS and points from the Arp et al. (2015) paper, as well as data from NSIDC. The conceptual diagram was adapted in PowerPoint, and the overall figure (as is the case for each of the six figures) was configured in lucidchart. 

# Figure 2

Similarly, Figure 2 was created in lucidchart, from a few Planet images, and an adapted conceptual diagram created in PowerPoint.

# Figure 3

Figure 3 can be recreated from the 2_CART_analysis.Rmd script, and the accuracy boxes and legend were created in lucidchart. 

# Figure 4

Script to reproduce Figure 4 is shown below, although access to Planet imagery is needed:

```{r load images}
require(jpeg)
library(sf)
library(ggplot2)

yolo_images = list.files("../data/images", pattern = ".jpg", full.names = TRUE)
yolo_labels = list.files("../data/labels", full.names = TRUE)
yolo_pred = list.files("../data/pred", full.names = TRUE)
```

```{r}
valid_bbox = list()
pred_bbox = list()
for(i in 1:length(yolo_images)){
  image_name = substr(list.files("../data/images")[i],1, 40)

  #then find matching tifs  
  label = list.files("../data/labels", pattern = image_name, full.names = TRUE)
  
  #ste0("data/out/", i, ".png"), width=8,height=8,units="in",res=300)
  img = readJPEG(yolo_images[i])
  plot(1:2,ty="n", axes = FALSE, frame.plot = FALSE)
  rasterImage(img,1,1,2,2)
  
  if(length(label>1)){
      valid_label = suppressWarnings(read.table(label))
       valid_bbox[[i]] = list()
      for(b in 1:nrow(valid_label)){
          #make the bounding box spatial
          valid_bbox[[i]][[b]] = st_polygon(list(rbind(c(valid_label$V2[b]-0.5*valid_label$V4[b]+1, (1-valid_label$V3[b])-0.5*valid_label$V5[b]+1),
                c(valid_label$V2[b]-0.5*valid_label$V4[b]+1, (1-valid_label$V3[b])+0.5*valid_label$V5[b]+1),
                c(valid_label$V2[b]+0.5*valid_label$V4[b]+1, (1-valid_label$V3[b])+0.5*valid_label$V5[b]+1),
                c(valid_label$V2[b]+0.5*valid_label$V4[b]+1, (1-valid_label$V3[b])-0.5*valid_label$V5[b]+1),
                c(valid_label$V2[b]-0.5*valid_label$V4[b]+1, (1-valid_label$V3[b])-0.5*valid_label$V5[b]+1))))
          plot(valid_bbox[[i]][[b]], lwd = 5, add= TRUE, border= "yellow", col = rgb(red = 0, green = 0, blue = 1, alpha = 0))
      }
  }
  pred = list.files("../data/pred",pattern = image_name, full.names = TRUE)
  if(length(label)>0){
      pred_label = read.table(pred)
      pred_bbox[[i]] = list()
      conf = list()
      for(bb in 1:nrow(pred_label)){
          #make the bounding box spatial
          pred_bbox[[i]][[bb]] = st_polygon(list(rbind(c(pred_label$V2[bb]-0.5*pred_label$V4[bb]+1, (1-pred_label$V3[bb])-0.5*pred_label$V5[bb]+1),
                c(pred_label$V2[bb]-0.5*pred_label$V4[bb]+1, (1-pred_label$V3[bb])+0.5*pred_label$V5[bb]+1),
                c(pred_label$V2[bb]+0.5*pred_label$V4[bb]+1, (1-pred_label$V3[bb])+0.5*pred_label$V5[bb]+1),
                c(pred_label$V2[bb]+0.5*pred_label$V4[bb]+1, (1-pred_label$V3[bb])-0.5*pred_label$V5[bb]+1),
                c(pred_label$V2[bb]-0.5*pred_label$V4[bb]+1, (1-pred_label$V3[bb])-0.5*pred_label$V5[bb]+1))))
          conf[[bb]] = pred_label$V6[bb]
          plot(pred_bbox[[i]][[bb]], lwd = 5, lty =2, add= TRUE, border = "turquoise1", col = rgb(red = 0, green = 1, blue = 0, alpha = 0))
          #text(pred_label$V2[bb]+0.5*pred_label$V4[bb]+1.03, (1-pred_label$V3[bb])+0.5*pred_label$V5[bb]+1.03, paste0(round(as.numeric(conf[bb]),2)), col = "turquoise1", cex = 1.5)
  
      }
  }
  #dev.off()
  
}


```

# Figure 5

Each panel in Figure 5 was mapped in QGIS, with basemap imagery from Google Satellite or NaturalEarth, using the predicted YOLOv5 bounding boxes linked on the accompanying Zenodo repository. 

# Figure 6

Bar charts in Figure 6 can be recreated below:

```{r load packages}
library(sf)
library(dplyr)
library(ggplot2)
```

```{r load input data and summarize}
MERIT_all = read_sf("../data/out/MERIT_4326.shp")
MERIT_all = MERIT_all %>% mutate(continent = factor(substr(MERIT_all$COMID,1,1))) %>% 
  mutate(perm = factor(substr(MERIT_all$COMBO,1,1)))

beaded_MERIT = read_sf("../data/out/MERIT_possible_beads.shp")
beaded_MERIT = beaded_MERIT %>% mutate(continent = factor(substr(beaded_MERIT$COMID,1,1)))%>% mutate(perm = factor(substr(beaded_MERIT$COMBO,1,1)))

Arp_int = read_sf("../data/out/Arp_int.shp")
Arp_int = data.frame(Arp_int) %>% distinct(ID, .keep_all = TRUE)
Arp_int = Arp_int %>% mutate(perm = factor(substr(Arp_int$COMBO,1,1))) %>% 
  mutate(continent = factor(substr(Arp_int$COMID,1,1)))

```

```{r load YOLOv5 prediction boxes}
predicted_boxes = read_sf("../data/out/YOLOv5_prediction_boxes.shp")
predicted_boxes = predicted_boxes %>% 
  mutate(perm = substr(predicted_boxes$COMBO,1,1)) %>% 
  mutate(continent = factor(substr(predicted_boxes$COMID,1,1)))

Permafrost = c(as.character(Arp_int$perm), predicted_boxes$perm, 
               as.character(beaded_MERIT$perm), as.character(MERIT_all$perm))

Continent = c(as.character(c(rep(8,.31*1900), rep(3, .69*1900))),
              as.character(predicted_boxes$continent),
              as.character(beaded_MERIT$continent),
              as.character(MERIT_all$continent))

data_combined = data.frame(Permafrost = c(c(as.character(Arp_int$perm), 
                                            c(rep("C", 1525))), 
                                          predicted_boxes$perm, 
                                          as.character(beaded_MERIT$perm), 
                                          as.character(MERIT_all$perm)),
                          Continent = c(as.character(c(rep(8,.31*1900), 
                                                       rep(3, .69*1900))), 
                                        as.character(predicted_boxes$continent), 
                                        as.character(beaded_MERIT$continent),
                                        as.character(MERIT_all$continent)),
                          Legend = c(rep("Arp et al., 2015", 1900), 
                                     rep("Computer Vision", nrow(predicted_boxes)),
                                     rep("Classified", nrow(beaded_MERIT)), 
                                     rep("Pan-Arctic", nrow(MERIT_all))))

data_combined$Test = factor(factor(data_combined$Legend), 
                            levels = levels(factor(data_combined$Legend))[c(1,3,2,4)])

data_new = data_combined %>% 
  filter(Legend %in% c("Arp et al., 2015", "Computer Vision"))

library(ggsci)

ggplot(data = data_new, aes(x = Test, fill = Continent)) +
  geom_bar(stat = "count") + theme_bw() + coord_flip() + ylab("Catchment Count") +
  scale_fill_manual(
    name = 'Continent',
    values = c('2' = '#0073C2FF', '3' = '#EFC000FF', '8' = '#868686FF'),
    labels = c('Europe', 'Asia', 'North America')
  ) +
  theme(text = element_text(size = 12), axis.title.y = element_blank())

data_combined$Test = factor(factor(data_combined$Legend), levels = levels(factor(data_combined$Legend))[c(3, 2, 1, 4, 5)])

levels(factor(data_combined$Legend))

ggplot(data = data_new, aes(x = Test, fill = Permafrost)) + 
  geom_bar(stat ="count", position = "fill") +
  theme_bw() + coord_flip() + ylab("Count") +
  scale_fill_manual(
    name = 'Permafrost',
    values = c(
      'C' = 'lightslategrey',
      'D' = 'lightskyblue3',
      'S' = 'lightblue1',
      'I' = 'mintcream',
      'g' = 'turquoise1',
      'l' = 'aquamarine4',
      'o' = 'darkseagreen1',
      'r' = 'burlywood4'
    ),
    labels = c(
      'Continuous',
      'Discontinuous',
      'Semi-continuous',
      'Isolated Permafrost',
      'Glacier',
      'Inland Lakes',
      'Ocean/Inland Sea',
      'Highlands'
    )
  ) +
  theme(text = element_text(size = 12), axis.title.y = element_blank()) +
  ylab("Proportion of Total Catchments")
```
